PLANNER-387 SolverFactory.createEmpty()
(cherry picked from commit 1b2ff82b73157a46900869335d006d8f53d5f186)
