+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/solver/EmptySolverFactory.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2015 JBoss Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.core.impl.solver;
+
+import org.optaplanner.core.config.solver.SolverConfig;
+
+public class EmptySolverFactory extends AbstractSolverFactory {
+
+    public EmptySolverFactory() {
+        this(null);
+    }
+
+    public EmptySolverFactory(ClassLoader classLoader) {
+        super(classLoader);
+        solverConfig = new SolverConfig();
+    }
+
+}
